requirements.txt
pymupdf
python-docx
googletrans==4.0.0-rc1
streamlit
